% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_function.R
\name{school_function}
\alias{school_function}
\title{A school Function}
\usage{
school_function(love = TRUE)
}
\arguments{
\item{love}{Do you love school? Defaults to TRUE.}
}
\description{
This function allows you to express your love for school.
}
\examples{
school_function()
}
\keyword{data}
